Fixed bugs
Completed Translation Function:
 - category, color and product translation
Redesigned Data Architecture:
 - changed category and color column from string type to id type of mapping to each table(Effect: when customer changed category or color name, it changes also for every product)
Fixed Other Bugs

+++ DB/cart.sql
@@ -1,9 +1,12 @@
 CREATE TABLE cart (
     id SERIAL PRIMARY KEY,
-    customer_id INT NOT NULL,
-    product_id INT NOT NULL,
-    quantity INT NOT NULL DEFAULT 1 CHECK (quantity > 0),
+    customer_id INTEGER NOT NULL,
+    product_id INTEGER NOT NULL,
+    quantity NUMERIC(10,1) NOT NULL DEFAULT 1 CHECK (quantity > 0),
     added_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    CONSTRAINT fk_cart_customer FOREIGN KEY (customer_id) REFERENCES customer(id) ON DELETE CASCADE,
-    CONSTRAINT fk_cart_product FOREIGN KEY (product_id) REFERENCES product(id) ON DELETE CASCADE
+    CONSTRAINT fk_cart_customer FOREIGN KEY (customer_id)
+        REFERENCES customer(id) ON DELETE CASCADE,
+
+    CONSTRAINT fk_cart_product FOREIGN KEY (product_id)
+        REFERENCES product(id) ON DELETE CASCADE
 );
